requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
nats-py
docker